Move MutualLink.ts to app
